% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_sim.R
\name{df_sim}
\alias{df_sim}
\title{df_sim}
\usage{
df_sim(sim, r)
}
\arguments{
\item{sim}{is the list output of the simulation result from the predprey_sim function}
}
\description{
Converts the output output list from the predprey_sim function into simulation data into a single dataframe.
}
\author{
Elizabeth Hiroyasu
}

