% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_alpha_plot.R
\name{k_alpha_plot}
\alias{k_alpha_plot}
\title{k_alpha_plot}
\usage{
k_alpha_plot(df, ymaxK, clean = TRUE)
}
\arguments{
\item{df}{is the dataframe to feed into ggplot2, it is the subsetted dataframe from the df_sim function in the 'owls' package.}

\item{ymaxK}{is the max limit for the y-axis, it is different depending on whether we're plotting for gophers or voles}
}
\description{
Uses ggplot2 package to plot the output data from predprey_sim in black and white.
}
\author{
Elizabeth Hiroyasu & Casey O'Hara
}
