% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsplit.R
\name{lsplit}
\alias{lsplit}
\title{List split}
\usage{
lsplit(sim, n, r)
}
\arguments{
\item{sim}{is the list output of the simulation result from the predprey_sim function}

\item{n}{is a sequence identifying the matrices to be extracted from the sim data}

\item{r}{is the r values specified in the parameters}
}
\description{
Splits the output list from the predprey_sim function into a list of outputs based on the initial population size 
of the prey population
}
\author{
Elizabeth Hiroyasu
}
