% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matplot_sim.R
\name{matplot_sim}
\alias{matplot_sim}
\title{Plotting simulations}
\usage{
matplot_sim(times, sim, s, r)
}
\arguments{
\item{times}{is the time vector to be used for the x axis, this should be the same as the t parameter in the pred_prey function}

\item{sim}{is the list of simulations}

\item{s}{is a vector of r values, usually this is a sample of r values}
}
\description{
This function can be used to plot the predator and prey populations over time for all simulation runs.
}
\author{
Elizabeth Hiroyasu
}

