% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.df.R
\name{n.df}
\alias{n.df}
\title{Converting initial simulation data to dataframes.}
\usage{
n.df(sim, r)
}
\arguments{
\item{sim}{is the list output from the predprey_sim function}

\item{r}{is a vector of r-values used in the simulation}
}
\description{
This function converts the list data from the n.sim function to data frames for each initial value of N. These data 
frames can then be used to plot the output data.
}
\author{
Elizabeth Hiroyasu
}
