% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.sim.R
\name{n.sim}
\alias{n.sim}
\title{N sim function}
\usage{
n.sim(sim, N, P, r)
}
\arguments{
\item{sim}{is the output list from the function predprey_sim}

\item{N}{is a vector of initial prey population densities}

\item{P}{is a vector of prey densities}

\item{r}{is a vector of population growth rates}
}
\description{
Creates a list of the simulation output by different initial N values. Each element is a different initial N value.
}
\author{
Elizabeth Hiroyasu
}
