% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.table.R
\name{n.table}
\alias{n.table}
\title{n.table}
\usage{
n.table(df, times, K_prey, N, alphaP)
}
\arguments{
\item{df}{is the output from the function df_sim}

\item{times}{is the length of time over which the scenario was run}

\item{K_prey}{is the carrying capacity of the prey species}

\item{N}{is the prey abundance}

\item{alpha}{is the attack rate of the predator}
}
\description{
Creates a table of the ode output by initial abundance values.
}
