% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_eq.R
\name{N_eq}
\alias{N_eq}
\title{Equilibrium N value}
\usage{
N_eq(P, r, K_prey, D, k_max)
}
\arguments{
\item{state}{is the state variables N for the prey population and P for the predator population; measured in individuals per ha.}

\item{parameters}{is a vector of parameters including r (population growth rate), alpha (attack rate), beta (assimilation efficiency of the predator),
delta (death rate of the predator), K_prey (carrying capacity of the prey population), k_max (max feeding rate of the predator), and D (half saturation constant)}
}
\description{
Uses the Lotka-Volterra predator-prey differential equations with a type III functional response
to calculate populations of both predator and prey populations over a period of time.
}
\author{
Elizabeth Hiroyasu
}
