% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_diff.R
\name{percent_diff}
\alias{percent_diff}
\title{Calculating percent difference in population density}
\usage{
percent_diff(sim, r)
}
\arguments{
\item{sim}{is the list output from the predprey_sim function}

\item{r}{is a vector of r-values used in the simulation}
}
\description{
This function calculates the total percent change in populations for the specified time period.
}
\author{
Elizabeth Hiroyasu
}

