% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_plot.R
\name{percent_plot}
\alias{percent_plot}
\title{Plot percent difference}
\usage{
percent_plot(sim, r)
}
\arguments{
\item{sim}{is the output list from the function n.sim (which is modified from the predprey_sim output)}

\item{r}{is a vector of population growth rates}
}
\description{
Plots the percent change in prey by different levels of predator density
}
\author{
Elizabeth Hiroyasu
}
