% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{Plotting change in populations}
\usage{
plot_diff(diff, N, j)
}
\arguments{
\item{diff}{is a matrix of differences by varying state variables; it is the output from the percent_diff function}

\item{N}{is the vector of input N values}

\item{j}{is a number calling which N value is being expressed}
}
\description{
This function can be used to plot the total percent change in populations over time. The x-axis should be predator density 
and the y axis is percent change in population of the prey species.
}
\author{
Elizabeth Hiroyasu
}
