% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eq.R
\name{plot_eq}
\alias{plot_eq}
\title{plot_eq}
\usage{
plot_eq(df, P, i)
}
\arguments{
\item{df}{is a dataframe that contains variables P, r, D, K_prey, N, and Eq. Eq is the equilibrium N1 or N2 variable}

\item{P}{is a vector of predator densities, this is used in the title of the plots}

\item{i}{is an interator to print the title}
}
\description{
Uses ggplot2 package to plot the output data from predprey_sim. Data must be processed into a dataframe using the lsplit
and df_sim functions.
}
\author{
Elizabeth Hiroyasu
}
