% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_grid.R
\name{plot_n_grid}
\alias{plot_n_grid}
\title{plot_n_grid}
\usage{
plot_n_grid(df, r, N, i)
}
\arguments{
\item{df}{is the dataframe to feed into the plot_n function, it is the subsetted dataframe from the df_sim function in the 'owls' package.}

\item{r}{is the r values specified in the parameters, because this function subsets by r value.}

\item{N}{is the vector of initial N values set in the state variables.}

\item{i}{is the value of N to call to specify the title.}
}
\description{
Uses the gridExtra package to plot the output plots from the plot_n function on one grid.
}
\author{
Elizabeth Hiroyasu
}

