% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_prey.R
\name{pred_prey}
\alias{pred_prey}
\title{Type II Lotka Volterra Function}
\usage{
pred_prey(t, state, parameters)
}
\arguments{
\item{t}{is a sequence of numbers with a specified time step}

\item{state}{is the state variables N for the prey population and P for the predator population; measured in individuals per ha.}

\item{parameters}{is a vector of parameters including r (population growth rate), alpha (attack rate), beta (assimilation efficiency of the predator),
delta (death rate of the predator), K_prey (carrying capacity of the prey population), k_max (max feeding rate of the predator), and D (half saturation constant)}
}
\description{
Uses the Lotka-Volterra predator-prey differential equations with a type III functional response
to calculate populations of both predator and prey populations over a period of time.
}
\author{
Elizabeth Hiroyasu
}
