% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_eq.R
\name{N_eq}
\alias{N_eq}
\title{Equilibrium N value}
\usage{
N_eq(P, D)
}
\arguments{
\item{P}{is the scaled predator density (individuals/ha), here P represents (kP)/K_prey}

\item{D}{is the scaled half saturation constant (individuals/ha), here D represents (K_prey)/(alpha*h)}
}
\description{
Uses the Lotka-Volterra predator-prey differential equations with a type III functional response
to calculate populations of both predator and prey populations over a period of time.
}
\author{
Elizabeth Hiroyasu
}
