% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{Plotting change in populations}
\usage{
plot_diff(state, diff)
}
\arguments{
\item{state}{is the state variables N for the prey population and P for the predator population; 
measured in individuals per ha. For this function, we are usually interested in the response of the prey species to
varying predator densities.}

\item{diff}{is a matrix of differences by varying state variables; it is the output from the percent_diff function}
}
\description{
This function can be used to plot the total percent change in populations over time. The x-axis should be predator 
and the y axis is percent change in population of the prey species.
}
\author{
Elizabeth Hiroyasu
}

